pyproject.toml
[project]
name = "pr-test-oracle"
version = "0.1.0"
description = "AI-powered service that analyzes GitHub PRs and recommends which tests to run"
requires-python = ">=3.12"
dependencies = [
    "fastapi",
    "uvicorn",
    "httpx",
    "pydantic-settings",
    "python-simple-logger",
]

[project.scripts]
pr-test-oracle = "pr_test_oracle.main:run"

[project.optional-dependencies]
dev = ["pytest", "pytest-asyncio"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["src"]

[tool.hatch.build.targets.wheel]
packages = ["src/pr_test_oracle"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    "D",       # pydocstyle - skip docstring enforcement
    "COM812",  # trailing-comma (conflicts with formatter)
    "ISC001",  # single-line-implicit-string-concatenation (conflicts with formatter)
    "S603",    # subprocess-without-shell-equals-true (we use subprocess intentionally)
    "S607",    # start-process-with-partial-path (CLI binaries)
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["S101", "PLR2004", "ANN", "SLF001", "S105", "S106", "S108", "PT011", "ARG001", "PLC0415"]
